[ecore] when testing database equality , first compare buffers and if not equal compare db structure and values
